refactor(:camera): accept String instead of Object
The camera provides a way to turn the camera off while the
component is still mounted. It also allows customization of the
MediaTrackConstraints object passed to `getUserMedia`. This however
makes this prop:

* harder to understand
* more error prone (Safari breaks for certain inputs)
* needlessly complex (only switching facingMode is really required)

To preserve the main functionality and make it simpler to use,
camera now accepts one of the following string values:

* "rear"  : use rear camera
* "front" : use front camera
* "none"  : turn camera off

BREAKING CHANGE: `:camera` prop does not accept objects or boolean
anymore.